requirements.txt
# Core LLM Integration
ollama>=0.1.0

# Vector Database
chromadb>=0.4.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.60.0
loguru>=0.7.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0

# Testing (dev dependencies)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code Quality (dev dependencies)
black
flake8
mypy
isort
